requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
duckdb>=0.9.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Machine learning
lightgbm>=4.0.0
scikit-learn>=1.3.0

# AI/LLM integration
google-generativeai>=0.3.0

# Optimization
ortools>=9.7.0
networkx>=3.1.0

# Time and date handling
pytz>=2023.3
python-dateutil>=2.8.2

# Excel file processing
openpyxl>=3.1.0
xlrd>=2.0.1

# Configuration management
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.0.0

# HTTP requests for webhooks and API integrations
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Security and authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
slowapi>=0.1.9
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0